requirements.txt
langchain-aws
langgraph
pandas
matplotlib
python-dotenv
boto3